# IoTRaspberrypi/openapi.yaml
...
openapi: 3.1.0
info:
  title: IoT and IM Microservices API
  description: API specification for Instant Messaging (Telegram, Line) and IoT (ESP32, Raspberry Pi) microservices.
  version: 1.0.0
  contact:
    name: Rocer52
    email: batte9306@gmail.com
servers:
  - url: http://localhost:5000
    description: Telegram API server
  - url: http://localhost:5001
    description: Line API server
  - url: http://localhost:5010
    description: ESP32 Device server
  - url: http://localhost:5011
    description: Raspberry Pi Device server
paths:
...
  /Enable:
    get:
      summary: Enable an IoT device
      description: Enable a specific ESP32 or Raspberry Pi device by device_id.
      parameters:
        - name: device_id
          in: query
          required: true
          schema:
            type: string
            enum:
              - esp32_light_001
              - esp32_fan_002
              - raspberrypi_light_001
              - raspberrypi_fan_002
            example: "esp32_light_001"
        - name: chat_id
          in: query
          required: false
          schema:
            type: string
            example: "7890547742"
        - name: platform
          in: query
          required: false
          schema:
            type: string
            enum:
              - telegram
              - line
            default: telegram
            example: "telegram"
        - name: user_id
          in: query
          required: false
          schema:
            type: string
            example: "987654321"
        - name: username
          in: query
          required: false
          schema:
            type: string
            example: "user123"
        - name: bot_token
          in: query
          required: false
          schema:
            type: string
            example: "YOUR_BOT_TOKEN"
      responses:
        '200':
          description: Device enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '400':
          description: Missing or invalid device_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
  /Disable:
    get:
      summary: Disable an IoT device
      description: Disable a specific ESP32 or Raspberry Pi device by device_id.
      parameters:
        - name: device_id
          in: query
          required: true
          schema:
            type: string
            enum:
              - esp32_light_001
              - esp32_fan_002
              - raspberrypi_light_001
              - raspberrypi_fan_002
            example: "esp32_light_001"
        - name: chat_id
          in: query
          required: false
          schema:
            type: string
            example: "7890547742"
        - name: platform
          in: query
          required: false
          schema:
            type: string
            enum:
              - telegram
              - line
            default: telegram
            example: "telegram"
        - name: user_id
          in: query
          required: false
          schema:
            type: string
            example: "987654321"
        - name: username
          in: query
          required: false
          schema:
            type: string
            example: "user123"
        - name: bot_token
          in: query
          required: false
          schema:
            type: string
            example: "YOUR_BOT_TOKEN"
      responses:
        '200':
          description: Device disabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '400':
          description: Missing or invalid device_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
  /GetStatus:
    get:
      summary: Get status of an IoT device
      description: Retrieve the current status of a specific ESP32 or Raspberry Pi device.
      parameters:
        - name: device_id
          in: query
          required: true
          schema:
            type: string
            enum:
              - esp32_light_001
              - esp32_fan_002
              - raspberrypi_light_001
              - raspberrypi_fan_002
            example: "esp32_light_001"
        - name: chat_id
          in: query
          required: false
          schema:
            type: string
            example: "7890547742"
        - name: platform
          in: query
          required: false
          schema:
            type: string
            enum:
              - telegram
              - line
            default: telegram
            example: "telegram"
        - name: user_id
          in: query
          required: false
          schema:
            type: string
            example: "987654321"
        - name: username
          in: query
          required: false
          schema:
            type: string
            example: "user123"
        - name: bot_token
          in: query
          required: false
          schema:
            type: string
            example: "YOUR_BOT_TOKEN"
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '400':
          description: Missing or invalid device_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceErrorResponse'
components:
  schemas:
...
    DeviceResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Device enabled"
        state:
          type: string
          example: "on"
        device_id:
          type: string
          example: "esp32_light_001"
      required:
        - status
        - message
        - state
        - device_id
    DeviceErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Missing device_id parameter"
      required:
        - status
        - message